Seclude CLI commands specific utils into commands folder
